refactor: use brick to handle key events
